requirements.txt
pandas
scikit-learn
seaborn
lightgbm
agno
python-dotenv
google-generativeai
